Manually copied wanted changes from bugfix/atm-geojson-performance re-added error handling that was removed by accident
