+++ Assets/Scripts/GeoJsonParser.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 15fc0f700fcb74e9890ebfca4e7b3f6a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
