+++ Assets/Scripts/Layers/Adapters/DataTypeAdapters/GeoJSONImportAdapter.cs
@@ -69,7 +69,7 @@ namespace Netherlands3D.Twin
             newLayer.Name = geoJsonLayerName;
             newLayer.gameObject.name = geoJsonLayerName;
             if (onErrorCallback != null)
-                newLayer.OnParseError.AddListener(onErrorCallback.Invoke);
+                newLayer.Parser.OnParseError.AddListener(onErrorCallback.Invoke);
 
             //GeoJSON layer+visual colors are set to random colors until user can pick colors in UI
             var randomLayerColor = Color.HSVToRGB(UnityEngine.Random.value, UnityEngine.Random.Range(0.5f, 1f), 1);
