+++ Assets/Scripts/Layers/LayerTypes/GeoJsonLayerGameObject.cs
@@ -2,30 +2,22 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.IO;
-using Newtonsoft.Json;
 using UnityEngine;
 using GeoJSON.Net;
 using GeoJSON.Net.CoordinateReferenceSystem;
 using GeoJSON.Net.Feature;
 using GeoJSON.Net.Geometry;
 using Netherlands3D.Coordinates;
-using SimpleJSON;
-using UnityEngine.Events;
 using Netherlands3D.Twin.Layers.Properties;
 using System.Linq;
-using Netherlands3D.LayerStyles;
 using Netherlands3D.Twin.Projects.ExtensionMethods;
-using UnityEngine.Networking;
-using Netherlands3D.SubObjects;
 
 namespace Netherlands3D.Twin.Layers
 {
     public class GeoJsonLayerGameObject : LayerGameObject, ILayerWithPropertyData
     {
-        public static float maxParseDuration = 0.01f;
-        
-        public GeoJSONObjectType Type { get; private set; }
-        public CRSBase CRS { get; private set; }
+        private GeoJsonParser parser = new GeoJsonParser(0.01f);
+        public GeoJsonParser Parser => parser;
 
         private GeoJSONPolygonLayer polygonFeaturesLayer;
         private GeoJSONLineLayer lineFeaturesLayer;
@@ -40,17 +32,38 @@ namespace Netherlands3D.Twin.Layers
         public int MaxFeatureVisualsPerFrame { get => maxFeatureVisualsPerFrame; set => maxFeatureVisualsPerFrame = value; }
 
         [Space]
-        public UnityEvent<string> OnParseError = new();
         protected LayerURLPropertyData urlPropertyData = new();
         public LayerPropertyData PropertyData => urlPropertyData;
 
+        private void Awake()
+        {
+            parser.OnFeatureParsed.AddListener(AddFeatureVisualisation);
+        }
+        
         protected override void Start()
         {
             base.Start();
-            if(urlPropertyData.Data.IsStoredInProject())
-                StartCoroutine(ParseGeoJSONStreamLocal(urlPropertyData.Data, 1000));
-            else if(urlPropertyData.Data.IsRemoteAsset())
-                StartCoroutine(ParseGeoJSONStreamRemote(urlPropertyData.Data, 1000));
+
+            if (urlPropertyData.Data.IsStoredInProject())
+            {
+                string path = Path.Combine(Application.persistentDataPath, urlPropertyData.Data.LocalPath.TrimStart('/', '\\'));
+                StartCoroutine(parser.ParseGeoJSONLocal(path));
+            }
+            else if (urlPropertyData.Data.IsRemoteAsset())
+            {
+                StartCoroutine(parser.ParseGeoJSONStreamRemote(urlPropertyData.Data));
+            }
+        }
+
+        private void OnDestroy()
+        {
+            parser.OnFeatureParsed.RemoveListener(AddFeatureVisualisation);
+        }
+
+        public void AddFeatureVisualisation(Feature feature)
+        {
+            VisualizeFeature(feature);
+            ProcessFeatureMapping(feature);
         }
 
         /// <summary>
@@ -79,159 +92,6 @@ namespace Netherlands3D.Twin.Layers
 
             if (pointFeaturesLayer != null)
                 pointFeaturesLayer.RemoveFeaturesOutOfView();
-        }
-
-        public void AppendFeatureCollection(FeatureCollection featureCollection)
-        {
-            var collectionCRS = featureCollection.CRS;
-            //Determine if CRS is LinkedCRS or NamedCRS
-            if (collectionCRS is NamedCRS)
-            {
-                if (CoordinateSystems.FindCoordinateSystem((collectionCRS as NamedCRS).Properties["name"].ToString(), out var globalCoordinateSystem))
-                {
-                    CRS = collectionCRS as NamedCRS;
-                }
-            }
-            else if (collectionCRS is LinkedCRS)
-            {
-                Debug.LogError("Linked CRS parsing is currently not supported, using default CRS (WGS84) instead"); //todo: implement this
-                return;
-            }
-
-        
-            StartCoroutine(VisualizeQueue(featureCollection.Features));
-        }
-
-        IEnumerator VisualizeQueue(List<Feature> features)
-        {
-            for (int i = 0; i < features.Count; i++)
-            {
-                var feature = features[i];
-
-                //If a feature was not found, stop queue
-                if (feature == null)
-                    yield break;
-                
-                VisualizeFeature(feature);
-                ProcessFeatureMapping(feature);
-
-                if (i % MaxFeatureVisualsPerFrame == 0)
-                    yield return null;
-            }
-        }
-
-        private IEnumerator ParseGeoJSONStreamRemote(Uri uri, int maxParsesPerFrame = Int32.MaxValue)
-        {
-            //create LocalFile so we can use it in the ParseGeoJSONStream function
-            string url = uri.ToString();
-            var uwr = UnityWebRequest.Get(url);
-
-            yield return uwr.SendWebRequest();
-            if (uwr.result == UnityWebRequest.Result.Success)
-            {
-                var startFrame = Time.frameCount;
-                // Get the downloaded text
-                string jsonText = uwr.downloadHandler.text;
-                StringReader reader = new StringReader(jsonText);
-                JsonTextReader jsonReader = new JsonTextReader(reader);
-                JsonSerializer serializer = new JsonSerializer();
-                serializer.Error += OnSerializerError;
-
-                FindTypeAndCRS(jsonReader, serializer);
-
-                //reset position of reader
-                jsonReader.Close();
-                reader.Dispose(); 
-
-                reader = new StringReader(jsonText); // Reset to start of JSON
-                jsonReader = new JsonTextReader(reader);
-
-                while (jsonReader.Read())
-                {
-                    // Read features depending on type
-                    if (jsonReader.TokenType == JsonToken.PropertyName && IsAtFeaturesToken(jsonReader))
-                    {
-                        jsonReader.Read(); // Start array
-                        yield return ReadFeaturesArrayStream(jsonReader, serializer);
-                    }
-                }
-                jsonReader.Close();
-
-                var frameCount = Time.frameCount - startFrame;
-                if (frameCount == 0)
-                    yield return null; // if entire file was parsed in a single frame, we need to wait a frame to initialize UI to be able to set the color.
-            }
-            else
-            {
-                OnParseError.Invoke("Dit GeoJSON bestand kon niet worden ingeladen vanaf de URL.");
-            }
-        }
-
-        private IEnumerator ParseGeoJSONStreamLocal(Uri uri, int maxParsesPerFrame = Int32.MaxValue)
-        {
-            string path = Path.Combine(Application.persistentDataPath, uri.LocalPath.TrimStart('/', '\\')); 
-          
-            var startFrame = Time.frameCount;
-            var reader = new StreamReader(path);
-            var jsonReader = new JsonTextReader(reader);
-
-            JsonSerializer serializer = new JsonSerializer();
-            serializer.Error += OnSerializerError;
-
-            FindTypeAndCRS(jsonReader, serializer);
-
-            //reset position of reader
-            jsonReader.Close();
-            reader = new StreamReader(path);
-            jsonReader = new JsonTextReader(reader);
-
-            while (jsonReader.Read())
-            {
-                //read features depending on type
-                if (jsonReader.TokenType == JsonToken.PropertyName && IsAtFeaturesToken(jsonReader))
-                {
-                    jsonReader.Read(); //start array
-                    yield return ReadFeaturesArrayStream(jsonReader, serializer);
-                }
-            }
-
-            jsonReader.Close();
-
-            var frameCount = Time.frameCount - startFrame;
-            if (frameCount == 0)
-                yield return null; // if entire file was parsed in a single frame, we need to wait a frame to initialize UI to be able to set the color.
-        }
-
-        private void OnSerializerError(object sender, Newtonsoft.Json.Serialization.ErrorEventArgs args)
-        {
-            OnParseError.Invoke("Er was een probleem met het inladen van dit GeoJSON bestand:\n\n" + args.ErrorContext.Error.Message);
-        }
-
-        private IEnumerator ReadFeaturesArrayStream(JsonTextReader jsonReader, JsonSerializer serializer)
-        {
-            var features = new List<Feature>();
-            var startTime = Time.realtimeSinceStartup;
-
-            while (jsonReader.Read())
-            {
-                if (jsonReader.TokenType == JsonToken.EndArray)
-                {
-                    // end of feature array, stop parsing here
-                    break;
-                }
-
-                var feature = serializer.Deserialize<Feature>(jsonReader);
-                features.Add(feature);
-                VisualizeFeature(feature);
-                ProcessFeatureMapping(feature);
-
-                var parseDuration = Time.realtimeSinceStartup - startTime;
-                if (parseDuration > maxParseDuration)
-                {
-                    yield return null;
-                    startTime = Time.realtimeSinceStartup;
-                }
-            }
         }
 
         private void ProcessFeatureMapping(Feature feature)
@@ -416,7 +276,7 @@ namespace Netherlands3D.Twin.Layers
         
         private void VisualizeFeature(Feature feature)
         {
-            var originalCoordinateSystem = GetCoordinateSystem();
+            var originalCoordinateSystem = GeoJsonParser.GetCoordinateSystem(feature.CRS);
             switch (feature.Geometry.Type)
             {
                 case GeoJSONObjectType.MultiPolygon:
@@ -489,127 +349,5 @@ namespace Netherlands3D.Twin.Layers
 
             polygonFeaturesLayer.AddAndVisualizeFeature<MultiPolygon>(feature, originalCoordinateSystem);
         }
-
-        private CoordinateSystem GetCoordinateSystem()
-        {
-            var coordinateSystem = CoordinateSystem.CRS84;
-
-            if (CRS is NamedCRS)
-            {
-                if (CoordinateSystems.FindCoordinateSystem((CRS as NamedCRS).Properties["name"].ToString(), out var globalCoordinateSystem))
-                {
-                    coordinateSystem = globalCoordinateSystem;
-                }
-            }
-            else if (CRS is LinkedCRS)
-            {
-                Debug.LogError("Linked CRS parsing is currently not supported, using default CRS (WGS84) instead"); //todo: implement this
-            }
-
-            return coordinateSystem;
-        }
-
-        private void FindTypeAndCRS(JsonTextReader reader, JsonSerializer serializer)
-        {
-            //reader must be at 0 for this to work properly
-            while (reader.TokenType != JsonToken.PropertyName)
-            {
-                reader.Read(); // read until property name is found (highest depth in json hierarchy)
-            }
-
-            bool typeFound = false;
-            bool crsFound = false;
-            do //process the found object, and continue reading after processing is done
-            {
-                // Log the current token
-                Debug.Log("Token: " + reader.TokenType + " Value: " + reader.Value);
-
-                if (!IsAtTypeToken(reader) && !IsAtCRSToken(reader))
-                {
-                    reader.Skip(); //if the found token is is not "type" or "crs", skip this object
-                }
-
-                //read type
-                if (IsAtTypeToken(reader))
-                {
-                    ReadType(reader, serializer);
-                    typeFound = true;
-                }
-
-                //read crs
-                if (IsAtCRSToken(reader))
-                {
-                    ReadCRS(reader, serializer);
-                    crsFound = true;
-                }
-
-                if (typeFound && crsFound)
-                    return;
-            } while (reader.Read());
-        }
-
-        private void ReadCRS(JsonTextReader reader, JsonSerializer serializer)
-        {
-            //Default if no CRS object is specified
-            CRS = DefaultCRS.Instance;
-            reader.Read(); // go to start of CRS object
-
-            //we need to stay within our CRS object, because there can also be "type" and "name" tokens outside of the object, and entire CRS objects in features.
-            //we must not accidentally parse these objects as our main CRS object, but we do not know the type we should deserialize as. We will just cast to a string and parse the object again since this is not a big string.
-            var CRSObject = serializer.Deserialize(reader);
-            var CRSString = CRSObject.ToString();
-
-            if (CRSString.Contains("link"))
-            {
-                var node = JSONNode.Parse(CRSString); //.DeserializeObject<LinkedCRS>(test);
-                var href = node["properties"]["href"];
-                var type = node["properties"]["type"];
-
-                CRS = new LinkedCRS(href, type);
-            }
-            else if (CRSString.Contains("name"))
-            {
-                var node = JSONNode.Parse(CRSString); //.DeserializeObject<LinkedCRS>(test);
-                var name = node["properties"]["name"];
-
-                CRS = new NamedCRS(name);
-            }
-        }
-
-        private void ReadType(JsonTextReader reader, JsonSerializer serializer)
-        {
-            if (reader.TokenType == JsonToken.PropertyName && IsAtTypeToken(reader))
-            {
-                reader.Read(); //read the value of the Type Object
-                Type = serializer.Deserialize<GeoJSONObjectType>(reader);
-                // print("parsed type: " + Type);
-            }
-        }
-
-
-        private static bool IsAtTypeToken(JsonTextReader reader)
-        {
-            if(reader.TokenType != JsonToken.PropertyName)
-                    return false;
-
-            return reader.Value.ToString().ToLower() == "type";
-        }
-
-        private static bool IsAtCRSToken(JsonTextReader reader)
-        {
-            if(reader.TokenType != JsonToken.PropertyName)
-                return false;
-
-            return reader.Value.ToString().ToLower() == "crs";
-        }
-
-        private static bool IsAtFeaturesToken(JsonTextReader reader)
-        {
-            if(reader.TokenType != JsonToken.PropertyName)
-                return false;
-
-            return reader.Value.ToString().ToLower() == "features";
-        }
-
     }
 }
